src/requirements.txt
flask==2.1.0
flask-sqlalchemy==2.5.1
mysqlclient==2.1.0
python-dotenv==0.20.0
gunicorn==20.1.0
# Dependências de teste
pytest==7.1.1
pytest-flask==1.2.0
coverage==6.3.2
